requirements.txt
aiohttp
aiofiles
numpy
pillow==9.5.0
asyncstdlib
apscheduler
ffmpeg
googletrans
yt-dlp
httpx[http2]>=0.13.3,<1.0.0
lyricsgenius
meval
git+https://github.com/KurimuzonAkuma/pyrogram.git
git+https://github.com/yt-dlp/yt-dlp.git
pytz
py-tgcalls
pyromod
requests
redis
speedtest-cli
SpeechRecognition
tgcrypto
telegraph
wget
youtube-search
youtube-search-python
